requirements.txt
streamlit>=1.28.0,<2.0.0
pdfplumber>=0.10.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0
pandas>=2.0.0,<3.0.0
plotly>=5.17.0,<6.0.0
python-dateutil>=2.8.0,<3.0.0
